JBIDE-18731 - fix weird packaging of archives core
JBIDE-18731 - delete deprecated class DirectoryScanner in package org.jboss.ide.eclipse.archives.core.asf

Cleanup as per fred
