+++ archives/plugins/org.jboss.ide.eclipse.archives.core/build.properties
@@ -1,21 +1,15 @@
 output.. = bin/
 bin.includes = META-INF/,\
                plugin.xml,\
-               archivescore.jar,\
                LICENSE-truezip.txt,\
                log4j.xml,\
-               archivescore-eclipse.jar,\
                lib/truezip-6.6.jar,\
                plugin.properties,\
-               about.html
-source.archivescore.jar = src/main/,\
-                          src/xml/
+               about.html,\
+               .
 src.includes = LICENSE-truezip.txt,\
                plugin.xml,\
                META-INF/,\
                src/
-output.archivescore.jar = bin/
-jars.compile.order = archivescore.jar,\
-                     archivescore-eclipse.jar
-source.archivescore-eclipse.jar = src/eclipse/
-output.archivescore-eclipse.jar = bin/
+source.. = src/main/,\
+           src/xml/
