+++ archives/tests/org.jboss.ide.eclipse.archives.test/src/org/jboss/ide/eclipse/archives/test/ArchivesTestSuite.java
@@ -13,7 +13,6 @@ package org.jboss.ide.eclipse.archives.test;
 import junit.framework.Test;
 import junit.framework.TestSuite;
 
-import org.jboss.ide.eclipse.archives.test.core.ArchivesCoreTest;
 import org.jboss.ide.eclipse.archives.test.core.ant.SimpleAntTest;
 import org.jboss.ide.eclipse.archives.test.model.DirectoryScannerModelTest;
 import org.jboss.ide.eclipse.archives.test.model.DirectoryScannerRegexTest;
@@ -38,7 +37,6 @@ import org.jboss.ide.eclipse.archives.test.util.TruezipUtilTest;
 public class ArchivesTestSuite extends TestSuite {
     public static Test suite() { 
         TestSuite suite = new TestSuite("Archives Tests");
-        suite.addTestSuite(ArchivesCoreTest.class);
         suite.addTestSuite(XBMarshallTest.class);
         suite.addTestSuite(XBUnmarshallTest.class);
         suite.addTestSuite(ReadWriteTest.class);
